# openapi-specification.yaml
openapi: 3.0.0
info:
  description: This is a description.
  version: 1.0.0
  title: Portfolio Tracker (MW212)
...
paths:
...
  /transactions/{transactionId}:
    get:
      tags:
        - transaction
      summary: Get transaction by user and id
      parameters:
        - name: userId
          in: path
          description: "The users id"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
    put:
      tags:
        - transaction
      summary: Update user
      description: This can only be done by the logged in user.
      parameters:
        - name: userId
          in: path
          description: "The users id"
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        description: Updated user object
        required: true
      responses:
        "400":
          description: Invalid user supplied
        "404":
          description: User not found
    delete:
      tags:
        - transaction
      summary: Delete user
      description: This can only be done by the logged in user.
      parameters:
        - name: userId
          in: path
          description: "The users id"
          required: true
          schema:
            type: string
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: User not found
...
servers:
  - url: https://petstore.swagger.io/v1
  - url: http://petstore.swagger.io/v1
components:
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: GUID
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
...